Restructured flight card
Extracted flight card into multiple class files and updatde directory structure
