+++ client/src/components/Dashboard/flights/flights.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from "axios";
-import { FlightCard } from "./viewFlights";
+import { FlightCard } from "./viewFlights/flightCard";
+import "./flights.css"
 
 
 const Flights = () => {
